Update README.md
In this project we have vividly studied and discuss about the deadliest and largest earthquakes since 1950. We have studied about the adverse side-effects of earthquakes like Tsunamis and casualties of those earthquakes.
We have discussed about the effects of these earthquakes on different countries using different ML techniques.
We suggested measures to be taken by different countries to counter earthquakes and other seismic activities.
We have also talked about the effects of Solar Flux Density and Solar Flares on megathrust earthquakes.